UIU-775-add-create-buttons-to-actions-menu add index.js; rename folder
